% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{add_dateline}
\alias{add_dateline}
\title{Add Date Line to Dataframe}
\usage{
add_dateline(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing incident date information}
}
\value{
A dataframe with added 'month_name', 'date_text', and 'incident_line' columns
}
\description{
Add Date Line to Dataframe
}
