% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{add_dateline_es}
\alias{add_dateline_es}
\title{Add Spanish Date Line to Dataframe}
\usage{
add_dateline_es(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing incident date information}
}
\value{
A dataframe with added 'month_name_es', 'date_text_es', and 'incident_line_es' columns
}
\description{
Add Spanish Date Line to Dataframe
}
