% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-responsibility-dated.R
\name{add_days_between}
\alias{add_days_between}
\title{Add a counter for days between events to event date tables}
\usage{
add_days_between(event_date_table)
}
\arguments{
\item{event_date_table}{as output by event_date_range_table}
}
\value{
A modified event date table with a variable `days_since_last`
  indicating the number of days since the `prior_event`
}
\description{
Add a counter for days between events to event date tables
}
\examples{
event_date_range_table(deaths_aug24) \%>\% add_days_between()
}
