% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-lookup.R
\name{add_id_for_event}
\alias{add_id_for_event}
\title{Add an unique identifier for event (id_event) column to a dataset}
\usage{
add_id_for_event(dataset, event_table = event.status, overwrite = FALSE)
}
\arguments{
\item{dataset}{A dataframe with a column containing the name of the event
in `event_title`}

\item{event_table}{A data frame containing `event_title`
and `id_event`}

\item{overwrite}{A boolean indicating whether to overwrite an existing
`id_event` variable}
}
\value{
A dataframe with a new column `id_event` containing the unique
  event ID, placed after the `event_title` column.
}
\description{
Add an unique identifier for event (id_event) column to a dataset
}
\examples{
add_id_for_event(deaths_aug24, event_status_aug24) \%>\%
  dplyr::select(event_title, id_event, id_indiv)
}
