% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-for-municipality.R
\name{add_id_for_municipality}
\alias{add_id_for_municipality}
\title{Add an unique identifier for municipality (id_muni) column to a dataset}
\usage{
add_id_for_municipality(
  dataset,
  id_variable_name = "id_muni",
  muni_table = anexo_municipios,
  overwrite = FALSE
)
}
\arguments{
\item{dataset}{A dataframe with a column containing the name of the
municipality, named `municipality`.}

\item{id_variable_name}{The name of the new column to be created. Default is
`id_muni`.}

\item{muni_table}{A data frame containing the municipality list table.}

\item{overwrite}{A boolean indicating whether to overwrite an existing
column.}
}
\value{
A dataframe with a new column `id_muni` containing the unique
  identifier for the municipality, placed before the `municipality` column.
}
\description{
Add an unique identifier for municipality (id_muni) column to a dataset
}
\examples{
id_for_municipality("Villa Ricardo Mugia - Icla", muni_list_table = muni_gb2014_lookup)
id_for_municipality_2("Villa Ricardo Mugia - Icla", "Chuquisaca")
}
