% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-for-municipality.R
\name{add_muni_gb2014_from_id}
\alias{add_muni_gb2014_from_id}
\title{Add Municipality GB2014 Code from ID}
\usage{
add_muni_gb2014_from_id(dataframe, muni_table = muni_id_lookup_table)
}
\arguments{
\item{dataframe}{A dataframe containing a column `id_muni`.}

\item{muni_table}{A lookup table containing `id_muni` and `muni_gb2014` columns.}
}
\value{
A dataframe with an additional column `muni_gb2014` containing the corresponding GB2014 codes.
}
\description{
This function adds a column `muni_gb2014` to a dataframe based on the `id_muni` column.
It uses a lookup table to map `id_muni` to `muni_gb2014`.
}
