% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{add_nameline}
\alias{add_nameline}
\title{Add Name Line to Dataframe}
\usage{
add_nameline(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing deceased person information}
}
\value{
A dataframe with added 'dec_fullname' and 'name_line' columns
}
\description{
Add Name Line to Dataframe
}
