% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{add_nameline_es}
\alias{add_nameline_es}
\title{Add Spanish Name Line to Dataframe}
\usage{
add_nameline_es(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing deceased person information}
}
\value{
A dataframe with added 'dec_fullname' and 'name_line_es' columns
}
\description{
Add Spanish Name Line to Dataframe
}
