% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-presidency-column.R
\name{add_presidency_column}
\alias{add_presidency_column}
\title{Add Presidency Column to Dataframe}
\usage{
add_presidency_column(
  dataframe,
  variable,
  dest_var,
  lookup_table = presidency_name_table,
  source_var = "presidency",
  .location = "right"
)
}
\arguments{
\item{dataframe}{The dataframe to which the presidency column will be added}

\item{variable}{The column name in the dataframe to match against the lookup
table}

\item{dest_var}{The name of the new column to create}

\item{lookup_table}{The table containing presidency information (defaults to
presidency_name_table)}

\item{source_var}{The column name in the lookup table to use as values
(defaults to "presidency")}

\item{.location}{The location to place the new column: "right" of all data
(default), "replace", or "beside" the original}
}
\value{
The original dataframe with the presidency column added
}
\description{
This function adds a column from the presidency lookup table to a dataframe
based on matching a specified key column.
}
\examples{
pres_freq_table <- count(assign_presidency_levels(deaths_aug24), pres_admin)
add_presidency_column(pres_freq_table, pres_admin, dest_var= "presidency_surnames")
add_presidency_column(pres_freq_table, pres_admin, dest_var= "presidency_surnames",
  .location="replace")
add_presidency_column(pres_freq_table, pres_admin, dest_var= "id_presidency",
  .location="beside")
add_presidency_column(pres_freq_table, pres_admin, dest_var= "presidency_year_es",
  .location="left")
add_presidency_column(pres_freq_table, pres_admin, dest_var= "presidency_initials_num",
  .location="right")
}
