% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{add_simple_label_es}
\alias{add_simple_label_es}
\title{Add Simple Label in Spanish to Dataframe}
\usage{
add_simple_label_es(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing event and deceased person information}
}
\value{
A dataframe with an added 'tooltip_text_es' column in Spanish
}
\description{
Add Simple Label in Spanish to Dataframe
}
\details{
This function creates a simple HTML-formatted label in Spanish for each row in the dataframe.
It includes information such as event title, deceased person's name, affiliation,
state responsibility, intentionality, cause of death, and incident date, all in Spanish.

The function first checks and repairs 'name_line' and 'incident_line' if they're missing.
}
