% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{affiliations}
\alias{affiliations}
\title{Helper variable for categories of affiliation}
\format{
A list with members `affiliations$levels` (a list of affiliation
categories in English), `affiliations$levels_es` (a list of affiliation
categories in Spanish), and `affiliations$colors` (a list of
affiliation categories with colors in English).
}
\usage{
affiliations
}
\description{
Helper variable for categories of affiliation
}
\keyword{datasets}
