% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{assign_dec_affiliation_levels}
\alias{assign_dec_affiliation_levels}
\alias{assign_perp_affiliation_levels}
\title{Assign Levels for Affiliation}
\usage{
assign_dec_affiliation_levels(dataframe, na.level = "Unknown")

assign_perp_affiliation_levels(dataframe, na.level = "Unknown")
}
\arguments{
\item{dataframe}{A dataframe containing a the relevant affiliation column}

\item{na.level}{The name of the level to which NA is assigned}
}
\value{
A dataframe with the affiliation column turned into an
  ordered factor
}
\description{
These functions factor the variable `dec_affiliation` and `perp_affiliation`.
}
\examples{
assign_dec_affiliation_levels(deaths_aug24)
assign_perp_affiliation_levels(deaths_aug24)
}
