% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{assign_presidency_levels}
\alias{assign_presidency_levels}
\title{Assign Presidency Levels}
\usage{
assign_presidency_levels(dataframe)
}
\arguments{
\item{dataframe}{A dataframe that may contain a 'pres_admin' column}
}
\value{
A dataframe with 'pres_admin' column as a factor with levels from president$levels
}
\description{
Assign Presidency Levels
}
\examples{
df <- data.frame(pres_admin = c("Obama", "Trump", "Biden"))
assign_presidency_levels(df)
}
