% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{assign_state_perpetrator_levels}
\alias{assign_state_perpetrator_levels}
\title{Assign State Perpetrator Levels}
\usage{
assign_state_perpetrator_levels(dataframe, simplify = FALSE)
}
\arguments{
\item{dataframe}{A dataframe containing a 'state_perpetrator' column}

\item{simplify}{Logical, whether to simplify the levels (default: FALSE)}
}
\value{
A dataframe with modified 'state_perpetrator' column
}
\description{
Assign State Perpetrator Levels
}
\examples{
df <- data.frame(state_perpetrator = c("yes", "no", NA, "LIKELY YES"))
assign_state_perpetrator_levels(df)
assign_state_perpetrator_levels(df, simplify = TRUE)
}
