% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{assign_state_responsibility_levels}
\alias{assign_state_responsibility_levels}
\title{Assign State Responsibility Levels}
\usage{
assign_state_responsibility_levels(dataframe, simplify = FALSE)
}
\arguments{
\item{dataframe}{A dataframe containing a 'state_responsibility' column}

\item{simplify}{Logical, whether to simplify the levels (default: FALSE)}
}
\value{
A dataframe with modified 'state_responsibility' column
}
\description{
This function simplifies the variable `state_responsibility`,
in part by overriding its values if the death was a conflict accident
(state_responsibility set to "Accidental") or an incidental death,
according to `intentionality`. NA's are resolved to "Unknown".
}
\details{
If simplify is FALSE, this function takes the existing values of
state_responsibility and turns them each into factors.

If simplify is TRUE, this function collapse to one-word values:
Perpetrator, Involved, Victim, Separate, Unintentional, Unknown.
}
\examples{
df1 <- data.frame(state_responsibility = c("State perpetrator", "Unknown", "State involved"),
  intentionality = c("Intentional", "Direct", "Conflict Accident"))
assign_state_responsibility_levels(df1)
assign_state_responsibility_levels(df1, simplify = TRUE)
assign_state_responsibility_levels(deaths_aug24)
}
