% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{clean_pt_variables}
\alias{clean_pt_variables}
\title{Clean variable names in presidents table}
\usage{
clean_pt_variables(pt)
}
\arguments{
\item{pt}{A dataframe with named variables}
}
\value{
A modified version of the dataframe
}
\description{
Puts all names into lower snake_case, removes "frequency of" at start
and any parenthetical expression at the end. Changes spaces, hyphens,
slashes to underscore. Removes question marks.
}
