% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{col2hex}
\alias{col2hex}
\title{Convert color names to hex RGB strings}
\usage{
col2hex(cname)
}
\arguments{
\item{cname}{Color name(s)}
}
\value{
Character vector giving the hex color code translation of
the provided color names.
}
\description{
Convert color names to hex RGB strings
}
\examples{
col2hex("forestgreen")
}
\author{
Gregory R. Warnes
}
