% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-names.R
\name{combine_names}
\alias{combine_names}
\title{Combine First and Last Names}
\usage{
combine_names(dataframe, unknown_string = "?")
}
\arguments{
\item{dataframe}{A dataframe containing 'dec_firstname' and 'dec_surnames' columns}

\item{unknown_string}{String to replace NA values (default: "?")}
}
\value{
A dataframe with an additional 'name' column
}
\description{
Combine First and Last Names
}
\examples{
df <- data.frame(dec_firstname = c("John", NA), dec_surnames = c("Doe", "Smith"))
combine_names(df)
}
