% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-plot-functions.R
\name{complete_x_values}
\alias{complete_x_values}
\title{New helper function to add null/missing x values}
\usage{
complete_x_values(
  counts_df,
  x_var,
  fill_var,
  all_levels = NULL,
  .verbose = FALSE
)
}
\arguments{
\item{counts_df}{The counts dataframe from waffle_counts}

\item{x_var}{The variable to facet the waffle chart by (e.g., year).}

\item{fill_var}{The variable to color the waffle chart blocks by (e.g.,
state_responsibility).}

\item{all_levels}{Optional: All possible levels for the x variable (needed
for pres_admin)}

\item{.verbose}{Logical indicating whether to print debug information}
}
\value{
The counts dataframe with missing x values added
}
\description{
New helper function to add null/missing x values
}
\examples{
\donttest{
deaths <- assign_levels(deaths_aug24, "standard", .simplify=TRUE)  \%>\%
  dplyr::filter(year != 2012)  # Remove 2012
waffle_counts_df <- waffle_counts(deaths,
                                 x_var = year,
                                  fill_var = protest_domain,
                                  fill_var_description = lev$protest_domain)
waffle_counts_df_completed <- complete_x_values(waffle_counts_df,
                                                x_var = year,
                                                fill_var = protest_domain,
                                                all_levels = NULL,
                                                .verbose = FALSE)
}
}
