% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-range-tools.R
\name{count_ongoing_events}
\alias{count_ongoing_events}
\title{Count Ongoing Events on a Specific Date}
\usage{
count_ongoing_events(event_date_range_table, input_date)
}
\arguments{
\item{event_date_range_table}{A table of date ranges of events in the
format produced by event_date_range_table()}

\item{input_date}{A Date object specifying the date to check for ongoing events}
}
\value{
An integer representing the number of events ongoing on the input date
}
\description{
Count Ongoing Events on a Specific Date
}
\examples{
event_range_table <- event_date_range_table(deaths_aug24)
num_events <- event_range_table \%>\% count_ongoing_events(as.Date("1988-06-27"))
print(paste("Number of ongoing events:", num_events))

}
