% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-range-by.R
\name{count_range_by}
\alias{count_range_by}
\title{Produce a Range Summary Table for Deaths by Responsibility}
\usage{
count_range_by(
  deaths,
  deaths_unfiltered,
  by,
  drop_separate = FALSE,
  drop_extra = FALSE,
  complete = TRUE,
  blank_hi = TRUE,
  .disqualified = FALSE,
  .verbose = FALSE
)
}
\arguments{
\item{deaths}{A data frame detailing confirmed and filtered deaths}

\item{deaths_unfiltered}{A data frame detailing all reported deaths,
including unconfirmed ones}

\item{by}{A variable to summarize by}

\item{drop_separate}{A logical value indicating whether to exclude the
"separate from state" category}

\item{drop_extra}{A logical value indicating whether to exclude extra
columns (e.g., unconfirmed, collateral, non-conflict deaths)}

\item{complete}{When true, fill in unknown counts with zeroes. This
does not override the blank_hi parameter.}

\item{blank_hi}{A logical value indicating whether to replace high estimates
with NA if they are the same as the low estimates}

\item{.disqualified}{When TRUE, return the list of events that would only
be included if the unfiltered list of deaths is used.}

\item{.verbose}{When TRUE, message with the number of events in the
filtered and unfiltered datasets.}
}
\value{
A data table with low and high estimates for each category of
state responsibility. In the default state, these are the index variable
supplied by "by" followed by "n", "n_state_perp", "n_state_perp_hi",
"n_state_victim", "n_state_victim_hi". When drop_separate is FALSE,
"n_state_separate" and "n_state_separate_hi" are also returned. When
drop_extra is false, "n_unfiltered", "n_unconfirmed", "n_collateral",
and "n_nonconflict" are also returned.
}
\description{
Given a table of deaths `deaths` and its unfiltered version `deaths_unfiltered`,
this function produces a summary table according to a given variable `by`,
calculating the range (low and high estimates) of the number of deaths where
the state was a perpetrator, victim, or separate from the incident.
}
\details{
The low estimates are based on the filtered `deaths` table, while the high
estimates are derived from the unfiltered `deaths_unfiltered` table, which
includes additional categories such as unconfirmed deaths, collateral deaths,
and non-conflict deaths.
}
\examples{
deaths_aug24_filtered <- standard_filter(deaths_aug24)
deaths_aug24_unfiltered <- deaths_aug24
count_range_by(deaths_aug24_filtered, deaths_aug24_unfiltered, protest_domain)
count_range_by(deaths_aug24_filtered, deaths_aug24_unfiltered, department)
}
