% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{deaths_aug24}
\alias{deaths_aug24}
\title{Ultimate Consequences Deaths Table}
\format{
`deaths_aug24` is a large data table. A tibble [670 × 59]
(S3: tbl_df/tbl/data.frame)
}
\usage{
deaths_aug24
}
\description{
An internal copy of the Ultimate Consequences deaths table
(not for public release)
}
\details{
c("event_title", "unconfirmed", "year", "month", "day", "later_year",
"later_month", "later_day", "dec_firstname", "dec_surnames",
"id_indiv", "dec_age", "dec_alt_age", "dec_gender", "dec_ethnicity",
"dec_residence", "dec_nationality", "dec_affiliation", "dec_spec_affiliation",
"dec_title", "cause_death", "cause_medical", "munition", "weapon",
"victim_armed", "va_notes", "perp_category", "perp_group", "perp_firstname",
"perp_surname", "perp_pol_stalemate", "dec_pol_stalemate", "perp_gender",
"perp_affiliation", "intentionality", "place", "location_precision",
"address", "community", "municipality", "province", "department",
"certainty_level", "pres_admin", "protest_campaign", "protest_domain",
"pol_assassination", "state_perpetrator", "sp_notes", "state_responsibility",
"sr_notes", "navarro_2006_number", "narrative", "day_notes",
"month_notes", "year_notes", "later_day_notes", "later_month_notes",
"later_year_notes")
}
\keyword{datasets}
