% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-outcome-reactable.R
\name{deaths_column}
\alias{deaths_column}
\title{Create a Column Definition for Death Counts}
\usage{
deaths_column(maxWidth = 60, class = NULL, maxValue = 100, ...)
}
\arguments{
\item{maxWidth}{Maximum width of the column (default: 60)}

\item{class}{CSS class for the column (default: NULL)}

\item{maxValue}{Maximum value for scaling the background color (default: 100)}

\item{...}{Additional arguments passed to colDef}
}
\value{
A column definition for use in reactable
}
\description{
Provides the necessary colum definition settings for a
death count column, including shaiding.
}
