% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_raw_sep24}
\alias{deaths_raw_sep24}
\title{An internal copy of the Ultimate Consequences deaths table raw data
(not for public release)}
\format{
`deaths_aug24` is a large data table. A tibble [675 × 50]
(S3: tbl_df/tbl/data.frame)
}
\usage{
deaths_raw_sep24
}
\description{
An internal copy of the Ultimate Consequences deaths table raw data
(not for public release)
}
\details{
c("event_title", "unconfirmed", "year", "month", "day", "later_year",
"later_month", "later_day", "dec_firstname", "dec_surnames",
"id_indiv", "dec_age", "dec_alt_age", "dec_gender", "dec_ethnicity",
"dec_residence", "dec_nationality", "dec_affiliation", "dec_spec_affiliation",
"dec_title", "cause_death", "cause_medical", "munition", "weapon",
"victim_armed", "perp_category", "perp_group", "perp_firstname",
"perp_surname", "perp_pol_stalemate", "dec_pol_stalemate", "perp_gender",
"perp_affiliation", "intentionality", "place", "location_precision",
"address", "community", "municipality", "province", "department",
"certainty_level", "pres_admin", "protest_campaign", "protest_domain",
"pol_assassination", "state_perpetrator", "state_responsibility",
"navarro_2006_number", "narrative")
}
\keyword{datasets}
