% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{departments}
\alias{departments}
\title{Helper Variable for Levels of Department}
\format{
A list with members `departments$title` (the string
"Department"), `departments$levels` (an ordered list of
department levels, with unknown in English), `departments$levels`
(an ordered list of department levels, with unknown in Spanish), and
`departments$colors` (name and color pairs for these levels).
}
\usage{
departments
}
\description{
Helper Variable for Levels of Department
}
\keyword{datasets}
