% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayed-date-string.R, R/estimate-date.R
\name{displayed_date_string}
\alias{displayed_date_string}
\title{Produce an Displayed Date String for (Sometimes Incomplete) Dates}
\usage{
displayed_date_string(year, month, day)

displayed_date_string(year, month, day)
}
\arguments{
\item{year}{Numerical year}

\item{month}{Numerical month}

\item{day}{Numerical day}
}
\value{
A string describing the date.

A string describing the date.
}
\description{
This function produces a displayable string using partial information
about a date to convey to the viewer exactly how much we know about the
date. When even the year is unknown, it reads "Date Unknown".
When the day is missing, it tell us "May 2010". Otherwise
the date is expressed in this format: "21 March 2024".

This function produces a displayable string using partial information
about a date to convey to the viewer exactly how much we know about the
date. When even the year is unknown, it reads "Date Unknown".
When the day is missing, it tell us "May 2010". Otherwise
the date is expressed in this format: "21 March 2024".
}
\examples{
displayed_date_string(2015, NA, NA), "2015-06-30")
displayed_date_string(2015, 8, NA), "2015-08-15")
displayed_date_string(2015, 1, 11)
displayed_date_string(2015, NA, NA), "2015-06-30")
displayed_date_string(2015, 8, NA), "2015-08-15")
displayed_date_string(2015, 1, 11)
}
