% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename-tools-subset.R
\name{domain_filename}
\alias{domain_filename}
\title{Generate protest domain filename}
\usage{
domain_filename(protest_domain)
}
\arguments{
\item{protest_domain}{The protest domain name}
}
\value{
The corresponding filename for the domain page,
        relative to the root of the website.
}
\description{
Generate protest domain filename
}
\examples{
domain_filename("Rural land")
}
