% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename-tools-subset.R
\name{domain_filename_es}
\alias{domain_filename_es}
\title{Generate Spanish domain filename}
\usage{
domain_filename_es(protest_domain)
}
\arguments{
\item{protest_domain}{The protest domain name in Spanish}
}
\value{
The corresponding filename for the Spanish domain page,
  relative to the root of the website.
}
\description{
Generate Spanish domain filename
}
\examples{
domain_filename_es("Campesino")
}
