% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-date.R
\name{estimated_date}
\alias{estimated_date}
\title{Estimate date based on known data}
\usage{
estimated_date(year, month, day)
}
\arguments{
\item{year}{Year as an integer}

\item{month}{Month as an integer}

\item{day}{Day as an integer}
}
\value{
A date in YYYY-MM-DD form, expressed as a string or Date.
}
\description{
This project works with imperfect data representing the dates of lethal
events. But when we need to order all the events chronologically, we must
estimate the dates involved. These function changes known year, month, and/or
day variables into estimated dates, expressed either as a string by
estimated_date_string() or as a Date variable by estimated_date().
}
\details{
When the day is unknovn, it is estimated as the 15th. When the year is
unknovn the estimated date is June 30. When the year is also unknovn,
returns NA.
}
\examples{
estimated_date_string(2015, NA, NA)
estimated_date_string(2015, 8, NA)
(estimated_date(2015, 10, 11)
}
