% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-date.R
\name{estimated_date_string}
\alias{estimated_date_string}
\title{Produce an Estimated Date String for (Sometimes Incomplete) Dates}
\usage{
estimated_date_string(year, month, day)
}
\arguments{
\item{year}{Numerical year}

\item{month}{Numerical month}

\item{day}{Numerical day}
}
\value{
A numerical date string in "YYYY-MM-DD" format
}
\description{
This function produces dates using partial information to enable
sequential sorting by date even when information is incomplete.
It treats unknown dates within a year as June 30 and unknown dates
within a month as the 15th day of the month.
}
\examples{
estimated_date_string(2015, NA, NA)
estimated_date_string(2015, 8, NA)
estimated_date_string(2015, 1, 11)
}
