% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-counts-by.R
\name{event_counts_by}
\alias{event_counts_by}
\title{Produce Table Summarized by a Custom Set of Variables, Each with an Event Lists}
\usage{
event_counts_by(def, ..., newline_style = "html")
}
\arguments{
\item{def}{Dataframe to be processed.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> List of variables in tidy-select format.}

\item{newline_style}{Text format for separating the event lists}
}
\value{
A dataframe listing all possible combinations of the
    variables being analyzed, followed by an events column, which
    has a text list of names of events each followed by the number
    of deaths involved in parentheses.
}
\description{
Produce Table Summarized by a Custom Set of Variables, Each with an Event Lists
}
\examples{
deaths_aug24 \%>\% event_counts_by(protest_domain, protest_campaign) \%>\%
  dplyr::filter(total>2) \%>\% dplyr::arrange(desc(total)) \%>\% print(n=10)
deaths_aug24 \%>\% event_counts_by(protest_domain, department) \%>\%
  dplyr::filter(total > 1) \%>\% dplyr::arrange(department, desc(total)) \%>\% print(n=100)
deaths_aug24 \%>\% event_counts_by(department) \%>\%
  dplyr::arrange(desc(total)) \%>\% dplyr::select(department) \%>\%
  dplyr::pull()
deaths_aug24 \%>\% event_counts_by(state_perpetrator, munition) \%>\%
  dplyr::filter(state_perpetrator=="Yes") \%>\% dplyr::arrange(desc(total)) \%>\% print(n=40)
}
