% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-outcomes-table.R
\name{event_counts_table}
\alias{event_counts_table}
\title{Generate a Table of Event Counts}
\usage{
event_counts_table(deaths, deaths_unfiltered, ...)
}
\arguments{
\item{deaths}{A data frame containing filtered deaths data.}

\item{deaths_unfiltered}{A data frame containing unfiltered deaths data.}

\item{...}{Additional arguments passed to the `count_range_by` function.}
}
\value{
A table of event counts.
}
\description{
This function generates a table of event counts based on the provided deaths data.
}
\examples{
deaths_aug24_filtered <- standard_filter(deaths_aug24)
deaths_aug24_unfiltered <- deaths_aug24
event_counts_table(deaths_aug24_filtered, deaths_aug24_unfiltered)
}
