% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-responsibility-dated.R
\name{event_date_range_table}
\alias{event_date_range_table}
\title{Produce a table of first and last days of each event and their length}
\usage{
event_date_range_table(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
A tibble with 11 variables: event_title, date_first, date_last,
  length_in_days, year_first, month_first, day_first, date_text_first,
  year_last, month_last, day_last, date_text_last
}
\description{
Produce a table of first and last days of each event and their length
}
\examples{
event_date_range_table(deaths_aug24)
}
