% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-counts-by.R
\name{event_list_with_counts_by}
\alias{event_list_with_counts_by}
\title{Produce Table of Events by a Custom Set of Variables}
\usage{
event_list_with_counts_by(dataframe, ...)
}
\arguments{
\item{dataframe}{Dataframe to be processed.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> List of variables in tidy-select format.}
}
\value{
A table whose left columns show all of the possible value
    combinations for the variables being analyzed. This is followed
    by a `count` column with the number of deaths registered for
    each corresponding event. Each event appears on its own line.
}
\description{
Produce Table of Events by a Custom Set of Variables
}
\examples{
deaths_aug24 \%>\% event_list_with_counts_by(protest_domain, protest_campaign) \%>\%
  dplyr::filter(count > 5) \%>\% dplyr::arrange(dplyr::desc(count))
}
