% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-outcomes-table.R
\name{event_outcomes_table}
\alias{event_outcomes_table}
\title{Create a Table of Event Outcomes}
\usage{
event_outcomes_table(campaign_events, event.status)
}
\arguments{
\item{campaign_events}{A data frame containing campaign events data.}

\item{event.status}{A data frame containing event status information.}
}
\value{
A table of event outcomes.
}
\description{
This function joins relevant event status information with campaign events data
to create a table of event outcomes.
}
\examples{
deaths_aug24_filtered <- standard_filter(deaths_aug24)
deaths_aug24_unfiltered <- deaths_aug24
event_counts <- event_counts_table(deaths_aug24_filtered, deaths_aug24_unfiltered)
event_outcomes_table(event_counts, event_status_aug24)
}
