% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-responsibility-dated.R
\name{event_responsibility_summary_table}
\alias{event_responsibility_summary_table}
\title{Create an Event Responsibility Summary Table}
\usage{
event_responsibility_summary_table(dataframe)
}
\arguments{
\item{dataframe}{Data table containing year, month, day}
}
\value{
A summary table describing events, with the following variables:
  `event_title`, `date`, `year`,
  `protest_domain`, `pres_admin`,
  `n`, `n_state_perp`, `n_state_victim`, `n_state_separate`
}
\description{
Given a properly factored table of deaths -- one on which the
state_responsibility variable has been factored and "Perpetrator" is one of
factor levels -- this function creates a summary table, each of whose
lines describes an individual event, by name, number of deaths,
number of state-perpetrated deaths, number of state victim deaths, and
number of deaths separate from the state.
}
\details{
Each event will also have the following characteristics recorded:
date (of first death), year, protest_domain, and presidential administration.
}
\examples{
assign_state_responsibility_levels(deaths_aug24, simplify=TRUE) |>
  event_responsibility_summary_table()
}
