% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{event_status_aug24}
\alias{event_status_aug24}
\title{Ultimate Consequences Deaths Table}
\format{
`event_status_aug24` is a large data table. A tibble [211 × 21]
(S3: tbl_df/tbl/data.frame)
}
\usage{
event_status_aug24
}
\description{
An internal copy of the Ultimate Consequences deaths table
(not for public release)
}
\details{
c("event_title", "year", "id_event", "n", "n_state_perp", "n_state_perp_excl",
"n_state_perp_direct_excl", "n_state_victim", "n_name_problems",
"name_problem_expl", "n_blank_notes", "n_blank_highestsource",
"n_blank_certainty", "outcome", "outcome_summary", "has_narrative",
"quality_narrative", "has_uncertain_deaths", "expand_narrative",
"newspaper", "sourcing")
}
\keyword{datasets}
