% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-collapse-insensitive.R
\name{fct_collapse_insensitive}
\alias{fct_collapse_insensitive}
\title{Collapse factors in a case-insensititve manner}
\usage{
fct_collapse_insensitive(f, ...)
}
\arguments{
\item{f}{A factor to collapse.}

\item{...}{Named arguments where the names are the new levels and the
values are vectors of old levels to be collapsed into the new level.}
}
\description{
This function collapses factor levels in a case-insensitive manner,
acting as a wrapper around `forcats::fct_collapse()`.
}
\details{
Accordingly, the capitalization of the old values of the factors
can be flexible, but the new levels must be specified with the
capitalization as desired.
}
\examples{
df2 <- data.frame(
  id_numbers = 1:10,
  quality = c("yes", "LIKELY YES", "no", "Presumed yes", "likely no")
)
df2$quality <- factor(df2$quality, levels = c("yes", "LIKELY YES", "no",
  "Presumed yes", "likely no"))
df2$quality <- fct_collapse_insensitive(df2$quality, list(
 Yes = c("Yes", "Likely Yes", "Presumed Yes"),
 No = c("No", "Likely No", "Presumed No")
))
}
