% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{fully_annotate}
\alias{fully_annotate}
\title{Add Simple Label to Dataframe}
\usage{
fully_annotate(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing event and deceased person information}
}
\value{
A dataframe with an added 'tooltip_text' column
}
\description{
Add Simple Label to Dataframe
}
\details{
This function creates a simple HTML-formatted label for each row in the dataframe.
It includes information such as event title, deceased person's name, affiliation,
state responsibility, intentionality, cause of death, and incident date.

The function first checks and repairs 'name_line' and 'incident_line' if they're missing.
}
