% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{generate_examples_with_results}
\alias{generate_examples_with_results}
\title{This function takes a string of function calls and generates Roxygen
`@examples` or `testthat` code with the results of those calls.
It evaluates each function call and appends the result as a comment.}
\usage{
generate_examples_with_results(calls, format = "examples")
}
\arguments{
\item{calls}{A string containing function calls, each on a new line.}

\item{format}{A string indicating the format of the output.
"examples" for Roxygen `@examples` or "tests" for `testthat`.}
}
\description{
IMPORTANT: You should only use a set of calls where you have manually
verified the results as correct for the function.
}
\examples{
generate_examples_with_results("sqrt(4)", "examples") #  "# examples #' sqrt(4) # \"2\""
calls <- 'n_filter(dplyr::starwars, (species=="Droid") & height<120)'
cat(generate_examples_with_results(calls, format = "tests"))
cat(generate_examples_with_results(calls, format = "examples"))
}
