% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-lookup.R
\name{id_for_event}
\alias{id_for_event}
\title{Get ID of an event based on the name}
\usage{
id_for_event(title, event_table = event.status)
}
\arguments{
\item{title}{Name of the event}

\item{event_table}{A data frame containing `event_title`
and `id_event`}
}
\value{
ID of the event as listed in the `event_table`. If no match is found,
  returns an empty string.
}
\description{
Get ID of an event based on the name
}
\examples{
id_for_event("Education strike 1987", event_status_aug24)
id_for_event("Huayllani ROADBLOCK", event_status_aug24)
}
