% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-for-municipality.R
\name{id_for_municipality}
\alias{id_for_municipality}
\alias{id_for_municipality_2}
\title{Get the ID of a municipality based on its name}
\usage{
id_for_municipality(
  municipality_i,
  department_i = "",
  muni_list_table = anexo_municipios
)

id_for_municipality_2(
  municipality_i,
  department_i = "",
  muni_list_table = muni_id_lookup_table
)
}
\arguments{
\item{municipality_i}{Name of the municipality to search for.}

\item{department_i}{Optional name of the department to filter by.}

\item{muni_list_table}{A data frame containing the municipality list table.}
}
\value{
The unique identifier for the municipality as listed in the `code` =
  column of the `muni_list_table`. If no match is found, returns an empty string.
}
\description{
Returns the unique identifier for a municipality based on its name and
optionally the department. These identifiers are the codes used
in the Bolivian government databases, known as INE codes for the
Instituto Nacional de Estadística.
}
\details{
The function will warn the user if a municipality name without
department corresponds to multiple municipalities. In this case, the first
municipality found will be returned. (There are about a dozen such cases in Bolivia.)

`id_for_municipality_2` is a version of the function that handles lookup
  tables with lists of municipalities. It is relies on `muni_id_lookup_table`.
}
\examples{
 id_for_municipality("La Paz") # 020101
 id_for_municipality("La Paz", "La Paz") # 020101
 id_for_municipality("La Paz", "Cochabamba") # empty string
 id_for_municipality("Potosí", "Potosí") # 050101
 id_for_municipality("Potosi") # 050101
 id_for_municipality("San Ramón") # 071103
 id_for_municipality("San Ramón", "Santa Cruz") # 071103
 id_for_municipality("San Ramón", "Beni") # 080702

 id_for_municipality_2("zudanez") # "010301"
 id_for_municipality_2("villa montes") # "060303"
 id_for_municipality_2("villamontes") # "060303"
 id_for_municipality_2("San Ignacio") # "070301"
 id_for_municipality_2("San Ignacio", "Beni") # "080501"
}
