% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-lookup.R
\name{id_for_president}
\alias{id_for_president}
\title{Get the ID of a presidency based on the name}
\usage{
id_for_president(
  name,
  pres_table = presidency_name_table,
  index_var = "id_presidency"
)
}
\arguments{
\item{name}{Name of a president or presidency}

\item{pres_table}{A data frame containing the presidency name table}

\item{index_var}{The name of the column to search for the name}
}
\value{
The unique identifier for the presidency as listed in the `index_var`
  column of the `pres_table`. If no match is found, returns an empty string.
}
\description{
Get the ID of a presidency based on the name
}
\examples{
id_for_president("Evo Morales")
id_for_president("Hernán Siles zuazo")
id_for_president("Jeanine Áñez")
id_for_president("HB") # returns a warning for multiple options
}
