% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-table-hierarchical.R
\name{insert_rows_by_group_vars}
\alias{insert_rows_by_group_vars}
\title{Insert rows into a grouped table from an additional dataframe}
\usage{
insert_rows_by_group_vars(main_df, additional_df, variable1, variable2)
}
\arguments{
\item{main_df}{A data.frame. The primary table containing the main groups.}

\item{additional_df}{A data.frame. Rows to be inserted into `main_df` by
matching on `variable1`.}

\item{variable1}{Column name (unquoted) used as the grouping key (tidy-eval).}

\item{variable2}{Column name (unquoted) indicating the second-level variable
to be inserted (tidy-eval).}
}
\value{
A data.frame containing rows from `main_df` with rows from
  `additional_df` inserted by group. Rows from `additional_df` whose
  `variable1` values do not exist in `main_df` will appear as new groups.
}
\description{
Insert rows from `additional_df` into `main_df` by matching groups defined
by `variable1`. For each unique value of `variable1` in `main_df`, if a
corresponding row exists in `additional_df` it will be appended to that
group's rows. Any `variable1` values that appear only in `additional_df`
will be appended as new groups at the end of the returned data frame.

This function uses tidy-eval for `variable1` and `variable2` (unquoted
column names).
}
\examples{
df_main <- data.frame(group = c("A", "A", "B"),
             sub = c("x", "y", "z"), n = 1:3, stringsAsFactors = FALSE)
df_add  <- data.frame(group = c("A", "C"),
             sub = c("other", "other"), n_2 = c(10, 5), stringsAsFactors = FALSE)
insert_rows_by_group_vars(df_main, df_add, group, sub)

}
\seealso{
two_layer_frequency_table, two_layer_frequency_kable
}
