% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{lev}
\alias{lev}
\title{A list of helper variables}
\format{
A list of helper variables for the package. The list contains:
`lev$pres_admin` (a list of levels of presidential administration),
`lev$location_precision` (a list of levels of location precision),
`lev$state_resp` (a list of levels of state responsibility),
`lev$protest_domain` (a list of levels of protest domain),
`lev$department` (a list of levels of department), `lev$dec_affilation`
and `lev$perp_affilation` (a list of levels of affiliation).
These names correspond to columns in the `deaths_aug24` data frame.
}
\usage{
lev
}
\description{
A list of helper variables
}
\keyword{datasets}
