% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_precision}
\alias{location_precision}
\title{Helper variable for levels of location precision}
\format{
A list with members `location_precision$levels`, a small-to-large
list of levels in English: c("address", "poi_small", "intersection",
"block", "poi_large", "road", "community", "town", "rural_zone",
"municipality", "province", "region", "department")
}
\usage{
location_precision
}
\description{
Helper variable for levels of location precision
}
\keyword{datasets}
