% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{luminance}
\alias{luminance}
\title{Calculate Luminance of a Hex-Coded Color}
\usage{
luminance(hex = "#ffffff")
}
\arguments{
\item{hex}{Color expressed in hex code}
}
\value{
Luminance between 0 and 1
}
\description{
Calculate Luminance of a Hex-Coded Color
}
\examples{
luminance("#ffe0e0")
luminance("#ff3030")
luminance("#bb2020")
}
