% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-plot-functions.R
\name{make_waffle_chart_tall}
\alias{make_waffle_chart_tall}
\title{Create a tall waffle chart with facets}
\usage{
make_waffle_chart_tall(
  dataframe,
  x_var,
  fill_var,
  fill_var_description,
  n_columns = 5,
  complete_x = FALSE,
  lang = "en",
  .verbose = FALSE
)
}
\arguments{
\item{dataframe}{The input dataframe containing the data.}

\item{x_var}{The variable to facet the waffle chart by (e.g., year
or pres_admin). This will name the facets.}

\item{fill_var}{The variable to color the waffle chart blocks by (e.g.,
state_responsibility).}

\item{fill_var_description}{A list containing the title, levels, and colors
for the fill variable. Can be taken from our level description list, lev.}

\item{n_columns}{Number of rows for the facet wrap (displayed vertically).}

\item{complete_x}{Logical indicating whether to complete missing x values
with a single null block. Only implemented for year and pres_admin.}

\item{lang}{Language for the legend labels ("en" or "es"). Default is "en".
Only affects the legend and not the x-axis label, which should be
handled before calling the function.}

\item{.verbose}{Logical indicating whether to print debug information.}
}
\value{
A ggplot object representing the tall waffle chart.
}
\description{
Creates a tall-oriented waffle chart from an overall dataset, both calculating
the relevant counts and formatting the chart. Facets are arranged vertically.
}
\examples{
\dontrun{
deaths <- assign_levels(deaths_aug24, "standard", .simplify = TRUE)
make_waffle_chart_tall(deaths,
  x_var = pres_admin,
  fill_var = state_responsibility,
  fill_var_description = lev$state_responsibility,
  complete_x = TRUE, n_columns = 7)
 make_waffle_chart_tall(deaths,
   x_var = pres_admin, fill_var = state_responsibility,
   fill_var_description = lev$state_responsibility,
   complete_x = FALSE, n_columns = 6)
 make_waffle_chart_tall(deaths,
   x_var = pres_admin, fill_var = state_responsibility,
   fill_var_description = lev$state_responsibility,
   complete_x = FALSE, n_columns = 6, lang="es")
}
}
