% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{month.name.es}
\alias{month.name.es}
\title{Table of Month Names in Spanish}
\format{
A list with twelve lower-case names as members from "enero"
  to "diciembre"
}
\usage{
month.name.es
}
\description{
Table of Month Names in Spanish
}
\keyword{datasets}
