% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni-counts-by.R
\name{muni_counts_by}
\alias{muni_counts_by}
\title{Produce Table Summarized by a Custom Set of Variables, Each with an Event List}
\usage{
muni_counts_by(def, ..., count_muni = FALSE, newline_style = "html")
}
\arguments{
\item{def}{Dataframe to be processed.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> List of variables in tidy-select format.}

\item{newline_style}{Text format for separating the event lists}
}
\value{
A dataframe listing all possible combinations of the
    variables being analyzed, followed by an events column, which
    has a text list of names of events each followed by the number
    of deaths involved in parentheses.
}
\description{
Produce Table Summarized by a Custom Set of Variables, Each with an Event List
}
\examples{
muni_counts_by(deaths_aug24, event_title) \%>\% dplyr::arrange(desc(total))
muni_counts_by(deaths_aug24, event_title, newline_style="oneline")
muni_counts_by(deaths_aug24, event_title, newline_style="oneline") \%>\%
  dplyr::mutate(len=stringr::str_length(municipalities)) \%>\%
  dplyr::arrange(desc(total)) \%>\% print(n=25)
}
