% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{muni_gb2014_conversion}
\alias{muni_gb2014_conversion}
\title{Municipality Name Conversion Table for GB2014 Format}
\format{
A data frame with columns:
\describe{
  \item{id_muni}{Municipal ID (INE code)}
  \item{muni_gb2014}{Municipality name in GB2014 format}
  \item{muni_anexo}{Standard municipality name from anexo}
  \item{recode}{R code string for recoding municipality names}
}
}
\usage{
muni_gb2014_conversion
}
\description{
A lookup table for converting between standard municipality names
and GB2014 format municipality names in Bolivia.
}
\details{
This table contains only municipalities where the GB2014 format name
differs from the standard anexo name. It includes generated R code
strings that can be used to perform the name conversions automatically.
}
\keyword{datasets}
