% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{muni_gb2014_lookup}
\alias{muni_gb2014_lookup}
\title{Municipality Lookup Table in GB2014 Format}
\format{
A data frame with columns:
\describe{
  \item{id_muni}{Municipal ID (INE code)}
  \item{municipality}{Municipality name in GB2014 format}
  \item{department}{Department name}
}
}
\usage{
muni_gb2014_lookup
}
\description{
A simplified lookup table using GB2014 format municipality names
for municipality ID and department lookup.
}
\details{
This table provides a streamlined lookup mechanism using GB2014
format municipality names. It's derived from muni_id_lookup_table
and can be used when working with data that uses GB2014 naming
conventions.
}
\keyword{datasets}
