% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni-counts-by.R
\name{muni_list_with_counts_by}
\alias{muni_list_with_counts_by}
\title{Produce a Summary Table that lists municipalities}
\usage{
muni_list_with_counts_by(dataframe, ...)
}
\arguments{
\item{dataframe}{Dataframe to be processed.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> List of variables in tidy-select format.}
}
\value{
A table whose left columns show all of the possible value
    combinations for the variables being analyzed. This is followed
    by a `count` column with the number of deaths registered for
    each corresponding municipality Each event appears on its own line.
}
\description{
This summary table aggregates municipality and department into
a singular location variable `muni_text` and appends it to
the right side of a summary table by an arbitrary set of variables.
}
\examples{
muni_list_with_counts_by(deaths_aug24, event_title)
}
