% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-for-municipality.R
\name{municipality_vector_from_id}
\alias{municipality_vector_from_id}
\alias{municipality_name_from_id}
\alias{province_name_from_id}
\alias{department_name_from_id}
\title{Get Municipality Details from ID Vector}
\usage{
municipality_vector_from_id(id_muni, muni_list_table = anexo_municipios)

municipality_name_from_id(id_muni, muni_list_table = anexo_municipios)

province_name_from_id(id_muni, muni_list_table = anexo_municipios)

department_name_from_id(id_muni, muni_list_table = anexo_municipios)
}
\arguments{
\item{id_muni}{A character vector of municipality IDs to look up.}

\item{muni_list_table}{A lookup dataframe containing municipality information.}
}
\value{
A list with three components:
  - `municipality`: A character vector of municipality names.
  - `province`: A character vector of province names.
  - `department`: A character vector of department names.

A string with the name of the municipality, province, or department.
}
\description{
This function looks up municipality details (municipality, province, and department)
from a list of municipality IDs in the given `muni_list_table`.
}
\details{
The function performs a lookup by matching the `id_muni` with the `cod.mun` column in `muni_list_table`.
If an ID is not found, the corresponding entries in the output list will be `NA`.
}
\examples{
library(dplyr)
# Example `muni_list_table`
anexo_municipios <- tibble::tibble(
  cod.mun = c("010101", "010102", "010103"),
  municipality = c("Sucre", "Yotala", "Poroma"),
  province = c("Oropeza", "Oropeza", "Oropeza"),
  department = c("Chuquisaca", "Chuquisaca", "Chuquisaca")
)

# Lookup municipality details
ids <- c("010101", "010102", "010999")
result <- municipality_from_id_muni(ids, anexo_municipios)
print(result)

municipality_name_from_id("020101") # "La Paz"
municipality_name_from_id("050101") # "Potosí"
municipality_name_from_id("071103") # "San Ramón"
municipality_name_from_id("080702") # "San Ramón"
municipality_name_from_id("010301") # "Villa Zudáñez"
province_name_from_id("020101") # "Murillo"
province_name_from_id("050101") # "Tomás Frías"
province_name_from_id("071103") # "Ñuflo de Chaves"
province_name_from_id("080702") # "Mamoré"
province_name_from_id("010301") # "Zudáñez"
province_name_from_id("060303") # "Gran Chaco"
province_name_from_id("070301") # "Velasco"
province_name_from_id("080501") # "Moxos"
department_name_from_id("020101") # "La Paz"
department_name_from_id("050101") # "Potosí"
department_name_from_id("071103") # "Santa Cruz"
department_name_from_id("080702") # "Beni"
department_name_from_id("010301") # "Chuquisaca"
department_name_from_id("060303") # "Tarija"
department_name_from_id("070301") # "Santa Cruz"
department_name_from_id("080501") # "Beni"
}
