% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-categorized-by.R
\name{n_categorized_by}
\alias{n_categorized_by}
\alias{n_responsibility_by}
\title{Produce a Summary Table for Deaths by Responsibility and Major Domains}
\usage{
n_categorized_by(def, by, complete = FALSE, sp_binary = FALSE)

n_responsibility_by(def, by, complete = FALSE)
}
\arguments{
\item{def}{A data frame detailing deaths}

\item{by}{A variable to summarize by}

\item{complete}{A logical value stating whether to fill in zero values}

\item{sp_binary}{A logical value. If TRUE, partial values for state_perpetrator
(e.g., "Likely Yes") are counted. Only used in n_categorized_by}
}
\value{
A data table.
}
\description{
Given a table of deaths `def`, these functions produce a summary table according
to a given variable `by` calculating the number of deaths in a variety
of situations.
}
\details{
The simpler function `n_responsibility_by` considers only the overall total number,
and deaths that were perpetrated by the state, had state forces as victims, or
were separate from the state. It returns a five-column data table.

The more complex function `n_categorized_by` subdivides deaths according to these categories,
as well as their involvement in the "Coca" protest domain, in the two domains
involving Armed Actors ("Guerrilla" and "Paramilitary"). This results in
a sixteen-column data table, where deaths outside those domains are
called Ordinary deaths.
}
\examples{
n_categorized_by(deaths_aug24, pres_admin, complete=TRUE)
n_categorized_by(deaths_aug24, pres_admin, complete=TRUE)
n_responsibility_by(deaths_aug24, protest_domain, complete=TRUE)
}
