% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-filter.R
\name{n_filter}
\alias{n_filter}
\title{Calculate number of rows matching a filter}
\usage{
n_filter(dataset, ...)
}
\arguments{
\item{dataset}{A dataframe usable by dplyr}

\item{...}{Other conditions to be passed to dplyr::filter}
}
\value{
An integer count of the number of matching rows/observations
}
\description{
Calculate number of rows matching a filter
}
\examples{
n_filter(dplyr::starwars, (species=="Droid") & height<120)
}
