% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-filter.R
\name{n_municipality}
\alias{n_municipality}
\title{Count observations in a specific municipality}
\usage{
n_municipality(dataset, stated_municipality)
}
\arguments{
\item{dataset}{A dataframe containing a "municipality" column}

\item{stated_municipality}{The municipality to filter by}
}
\value{
An integer count of rows matching the specified municipality
}
\description{
Count observations in a specific municipality
}
\examples{
n_municipality(deaths_aug24, "Cochabamba")
}
