% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{n_observations_to_reach_share}
\alias{n_observations_to_reach_share}
\title{Calculate Number of Observations to Reach a Given Share}
\usage{
n_observations_to_reach_share(datavector, share = 0.5)
}
\arguments{
\item{datavector}{A numeric vector of values}

\item{share}{Numeric value between 0 and 1 representing the target share}
}
\value{
An integer representing the number of largest values needed to
  reach the target share
}
\description{
This function determines how many of the largest observations in a vector are needed to reach a specified share of the total sum.
}
\details{
The function sorts the input vector in descending order and iteratively checks how many of the largest values are needed to exceed the specified share.
}
