% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{perp_pal}
\alias{perp_pal}
\alias{sv_pal}
\alias{sep_pal}
\title{Color Palettes by Category for a Numerical Scale}
\usage{
perp_pal(...)

sv_pal(...)

sep_pal(...)
}
\arguments{
\item{...}{Parameters to be passed to the underlying palette}
}
\value{
A color along the palette range
}
\description{
A series of functions blending from a light grey to
category-specific colors: perp_pal() for state-perpetrated
deaths; sv_pal() for state-victim deaths, sep_pal() for
deaths separate from the state. Each passes `x`, a number from 0 to 1
to the underlying palette function.
}
\examples{
max_deaths <- 100
red_pal(80/max_deaths)
brown_pal(80/max_deaths)
bluegray_pal(80/max_deaths)
gray_pal(80/max_deaths)
}
