% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{presidency_name_table}
\alias{presidency_name_table}
\title{Helper variable for names for presidential administrations}
\format{
A tibble with multiple columns including unqiue identifiers in
  `presidency` (the levels of pres_admin in English) and `presidency_id`
  (a unique identifier in the format "p104"). `first_day` and `last_day`
  are dates indicating the term of offfice of each president.
}
\usage{
presidency_name_table
}
\description{
A look-up table for presidential administrations in Bolivia from
1964 to the present. Columns are as follows:
- presidency: Name of the presidential administration in English
- id_presidency: Unique identifier for the presidential administration
- presidency_year: Common name for the president, plus years in office
    in parentheses
- presidency_commonname: Common name for the president
- presidency_fullname: Full name of the president
- presidency_surnames: Surnames of the president
- presidency_year_es, presidency_commonname_es, presidency_fullname_es:
    As above but with text (besides the name) in Spanish
- presidency_initials: Initials of the president
- presidency_initials_num: Initials of the president with multiple
    terms distinguished by numbers and years of Junta labeled
- first_day: Date of the first day of the presidential term
- last_day: Date of the last day of the presidential term
}
\keyword{datasets}
