% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{protest_domains}
\alias{protest_domains}
\title{Helper Variable for Levels of Protest Domain}
\format{
A list with members `protest_domain$title` (the string "Protest
Domain"), `protest_domain$levels` (an ordered list of
state responsibility levels in English), `protest_domain$colors` (name
and color pairs for these levels in English).
}
\usage{
protest_domains
}
\description{
Helper Variable for Levels of Protest Domain
}
\keyword{datasets}
