% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasilogical-as-binary.R
\name{quasilogical_as_binary}
\alias{quasilogical_as_binary}
\title{Turn quasilogical variable into a binary one}
\usage{
quasilogical_as_binary(datacolumn)
}
\arguments{
\item{datacolumn}{The relevant variable in a dataframe}
}
\value{
Same data column with simplified values
}
\description{
We use "quasilogical" variables to document uncertainty about facts
("likely" means we lack certainty) and about overt claims ("Presumed"
denotes a logical inference like that civilians are unarmed). This
function flattens quasilogical variables into three options:
"Yes", "No" and "Disputed". The project
}
\examples{
quasilogical_as_binary(deaths_aug24$state_perpetrator)
}
