% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{red_pal}
\alias{red_pal}
\alias{brown_pal}
\alias{bluegray_pal}
\alias{gray_pal}
\title{Color Palettes for a Numerical Scale}
\usage{
red_pal(x)

brown_pal(x)

bluegray_pal(x)

gray_pal(x)
}
\arguments{
\item{x}{A number from 0 to 1}
}
\value{
A color along the palette range
}
\description{
A series of functions with different colors:
red_pal(), brown_pal(), bluegray_pal(), gray_pal().
}
\examples{
max_deaths <- 100
red_pal(80/max_deaths)
brown_pal(80/max_deaths)
bluegray_pal(80/max_deaths)
gray_pal(80/max_deaths)
}
