% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-for-municipality.R
\name{rename_anexo_columns}
\alias{rename_anexo_columns}
\title{Rename Columns in a Dataframe for Standardization}
\usage{
rename_anexo_columns(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing the columns to be standardized.}
}
\value{
A dataframe with standardized column names.
}
\description{
This function renames specific columns in a dataframe to standardize their names.
It ensures that the columns `municipality`, `province`, and `department` are present, renaming existing columns
(`municipio`, `provincia`, `departamento`) if necessary. Additionally, it renames any column starting with `codigo`
to `cod.mun`.
}
\details{
- The function checks for the presence of `municipality`, `province`, and `department`.
  If these columns are missing, it renames `municipio`, `provincia`, and `departamento` to their English equivalents.
- Any column name starting with `codigo` is renamed to `cod.mun`.
}
\examples{
# Example dataframe
df <- data.frame(
  municipio = c("La Paz", "Cochabamba"),
  provincia = c("Pedro Domingo Murillo", "Cercado"),
  departamento = c("La Paz", "Cochabamba"),
  codigo123 = c(101, 102)
)

# Rename columns
renamed_df <- rename_anexo_columns(df)
print(names(renamed_df)) # Should include "municipality", "province", "department", "cod.mun"

}
