% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-age.R
\name{render_age}
\alias{render_age}
\alias{render_age_es}
\title{Render a numerical age as text}
\usage{
render_age(age)

render_age_es(age)
}
\arguments{
\item{age}{A real number or fraction experssing age in years}
}
\value{
A text string stating the age using digits and words,
  providing only years for over 2 years old; years and months fro
  under two years old; and "less than 30 days" where appropriate.
}
\description{
Returns a partial numerical and partially spelled out age for
a person. `render_age` is in English and `render_age_es` is
in Spanish
}
\examples{
render_age(7)
render_age(2+5/12)
render_age_es(1+5/12)
render_age(8/12)
render_age_es(33/365)
render_age(28/365)
}
