% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{repair_incident_line}
\alias{repair_incident_line}
\title{Repair Incident Line in Dataframe}
\usage{
repair_incident_line(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing incident information}
}
\value{
A dataframe with a repaired 'incident_line' column
}
\description{
Repair Incident Line in Dataframe
}
