% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{repair_incident_line_es}
\alias{repair_incident_line_es}
\title{Repair Spanish Incident Line in Dataframe}
\usage{
repair_incident_line_es(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing incident information}
}
\value{
A dataframe with a repaired 'incident_line_es' column in Spanish
}
\description{
Repair Spanish Incident Line in Dataframe
}
