% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{repair_name_line}
\alias{repair_name_line}
\title{Repair Name Line in Dataframe}
\usage{
repair_name_line(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing deceased person information}
}
\value{
A dataframe with a repaired 'name_line' column
}
\description{
Repair Name Line in Dataframe
}
