% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-functions.R
\name{repair_name_line_es}
\alias{repair_name_line_es}
\title{Repair Spanish Name Line in Dataframe}
\usage{
repair_name_line_es(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing deceased person information}
}
\value{
A dataframe with a repaired 'name_line' column in Spanish
}
\description{
Repair Spanish Name Line in Dataframe
}
