% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{set_maximums}
\alias{set_maximums}
\title{Set Maximum Death Quantities}
\usage{
set_maximums(df.responsibility)
}
\arguments{
\item{df.responsibility}{A dataframe with n, n_state_victim, n_state_separate}
}
\value{
maxm: A list with three values $deaths, $sv_deaths, $sep_deaths
}
\description{
Set Maximum Death Quantities
}
\examples{
n_responsibility_by(deaths_aug24, department) \%>\% set_maximums()
}
