% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{share_of_largest_n}
\alias{share_of_largest_n}
\title{Calculate Share of Largest N Values}
\usage{
share_of_largest_n(datavector, rank = 1)
}
\arguments{
\item{datavector}{A numeric vector of values}

\item{rank}{Integer specifying the number of largest values to consider (default: 1)}
}
\value{
A numeric value representing the share of the sum of the largest n values compared to the total sum
}
\description{
This function calculates the share of the sum of the largest n values in
a vector compared to the total sum of all values in the vector.
}
\details{
The function sorts the input vector in descending order, sums the first n
values (where n is specified by the rank parameter), and divides this by
the total sum of the vector.
}
\examples{
share_of_largest_n(c(10, 5, 3, 2), rank = 1)
share_of_largest_n(c(10, 5, 3, 2), rank = 2)

}
