% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{standard_factoring_variables}
\alias{standard_factoring_variables}
\title{List of variables to be factored}
\format{
A list of variables to be factored in the data set. The list
contains:
  "pres_admin", "state_responsibility", "state_perpetrator",
  "protest_domain", "dec_affiliation", "perp_affiliation",
  "location_precision."
}
\usage{
standard_factoring_variables
}
\description{
List of variables to be factored
}
\keyword{datasets}
