% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-filter.R
\name{standard_filter}
\alias{standard_filter}
\alias{filter_out_unintentional}
\alias{filter_out_unconfirmed}
\title{Filter a Dataframe to Remove Unintentional and/or Unconfirmed Deaths}
\usage{
standard_filter(dataframe)

filter_out_unintentional(dataframe)

filter_out_unconfirmed(dataframe)
}
\arguments{
\item{dataframe}{A data frame containing information about deaths}
}
\value{
The input dataframe with the specified types of deaths removed
}
\description{
These functions filter a dataframe to remove deaths that are considered unintentional
or unconfirmed, based on the values in the `intentionality` and `unconfirmed` columns.
}
\details{
`standard_filter()` removes both unconfirmed deaths and deaths with "Collateral",
"Nonconflict Accident", or "Nonconflict" values in the `intentionality` column.
If either the `intentionality` or `unconfirmed` column is missing, a message is
displayed, but the function still filters based on the available column.

`filter_out_unintentional()` removes only deaths with "Collateral", "Nonconflict Accident",
or "Nonconflict" values in the `intentionality` column. If the `intentionality` column
is missing, a message is displayed and the original dataframe is returned.

`filter_out_unconfirmed()` removes only unconfirmed deaths based on the `unconfirmed`
column. If the `unconfirmed` column is missing, a message is displayed and the original
dataframe is returned.
}
\examples{
filtered_data <- standard_filter(deaths_aug24)
intentional_data <- filter_out_unintentional(deaths_aug24)
confirmed_data <- filter_out_unconfirmed(deaths_aug24)
}
