% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-data.R
\docType{data}
\name{state_resp}
\alias{state_resp}
\title{Helper variable for levels of state responsibility}
\format{
A list with members `state_resp$levels` (an ordered list of
state responsibility levels in English), `state_resp$colors` (name
and color pairs for these levels in English) and `state_resp$colors_es`,
the same pairs with Spanish names
}
\usage{
state_resp
}
\description{
Helper variable for levels of state responsibility
}
\keyword{datasets}
