% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-equivalent.R
\name{str_equivalent}
\alias{str_equivalent}
\title{Check if strings are equivalent (discarding accents and capitalization)}
\usage{
str_equivalent(x, y)
}
\arguments{
\item{x}{string}

\item{y}{string}
}
\value{
Boolean TRUE/FALSE indicating if the strings are equivalent
}
\description{
Check if strings are equivalent (discarding accents and capitalization)
}
\examples{
str_equivalent("Evo Morales", "evo morales")
str_equivalent("Carlos\u00a0Mesa", "Carlos Mesa")
}
