% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-helper.R
\name{string_to_listcase}
\alias{string_to_listcase}
\title{Convert String to List Case}
\usage{
string_to_listcase(string)
}
\arguments{
\item{string}{A character string to be converted}
}
\value{
A character string in list case format
}
\description{
List case here is defined as "Sentence case between commas,
With new capital letters only after, Each comma."
}
\examples{
string_to_listcase("hello, world")
}
