% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-counts-by.R
\name{top_values_string}
\alias{top_values_string}
\title{Get Top Values as a String}
\usage{
top_values_string(freq_table, n_values = 3, incl_counts = FALSE, by = "total")
}
\arguments{
\item{freq_table}{A data frame containing frequency data}

\item{n_values}{Integer. Number of top values to return. Default is 3.}

\item{incl_counts}{Logical. Whether to include counts in the output string. Default is FALSE.}

\item{by}{Name of the frequency column}
}
\value{
A string containing the top values, separated by commas
}
\description{
Get Top Values as a String
}
\examples{
freq_table <- data.frame(protest_domain = c("A", "B", "C"), total = c(10, 5, 3))
top_values_string(freq_table)
top_values_string(freq_table, n_values = 2, incl_counts = TRUE)
evco <- event_counts_by(deaths_aug24, protest_domain, count_events=TRUE)
top_values_string(evco, 5, incl_counts = TRUE, by="n_events")

}
