% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-counts-by.R
\name{truncate_event_list}
\alias{truncate_event_list}
\title{Process Event Count Strings to First N Events}
\usage{
truncate_event_list(
  dataframe,
  variable = "events",
  num_events = 4,
  sep_char = ","
)
}
\arguments{
\item{dataframe}{Dataframe to be processed.}

\item{variable}{Variable containing the event string
(`events` by default)}

\item{num_events}{Number of events to be kept in string}

\item{sep_char}{Character that separates the listed items}
}
\value{
A dataframe with the designated variable truncated.
}
\description{
Process Event Count Strings to First N Events
}
\examples{
deaths_aug24 \%>\% event_counts_by(department, newline_style="text") \%>\%
  truncate_event_list(num_events = 6) \%>\% dplyr::arrange(desc(total))
}
