% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni-counts-by.R
\name{truncate_muni_list}
\alias{truncate_muni_list}
\title{Process Municipality Count Strings to First N Municipalities}
\usage{
truncate_muni_list(
  dataframe,
  variable = "municipalities",
  num_muni = 4,
  sep_char = ";"
)
}
\arguments{
\item{dataframe}{Dataframe to be processed.}

\item{variable}{Variable containing the event string
(`municipalities` by default)}

\item{num_muni}{Number of events to be kept in string}

\item{sep_char}{Single-character symbol used to separate the list
(`;` by default)}
}
\value{
A dataframe with the designated variable dataed.
}
\description{
Process Municipality Count Strings to First N Municipalities
}
\examples{
deaths_aug24 \%>\% muni_counts_by(department, newline_style="text") \%>\%
  truncate_muni_list(num_muni = 3) \%>\% dplyr::arrange(desc(total))
}
