% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle-plot-functions.R
\name{waffle_counts}
\alias{waffle_counts}
\title{Helper function to create counts dataframe for waffle charts}
\usage{
waffle_counts(dataframe, x_var, fill_var, fill_var_description)
}
\arguments{
\item{dataframe}{The input dataframe containing the data.}

\item{x_var}{The variable to facet the waffle chart by (e.g., year).}

\item{fill_var}{The variable to color the waffle chart blocks by (e.g.,
state_responsibility).}

\item{fill_var_description}{A list containing the title, levels, and colors
for the fill variable. Can be taken from our level description list, lev.}
}
\value{
A dataframe with counts of occurrences for each combination of x_var
  and fill_var.
}
\description{
Helper function to create counts dataframe for waffle charts
}
\examples{
waffle_counts(dataframe = assign_state_responsibility_levels(deaths_aug24, simplify = TRUE),
 x_var = year, fill_var = state_responsibility,
              fill_var_description = lev$state_responsibility)
}
