% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-range-tools.R
\name{count_events_in_month}
\alias{count_events_in_month}
\title{Count Events Overlapping with a Specific Month}
\usage{
count_events_in_month(event_date_range_table, yearmonth)
}
\arguments{
\item{yearmonth}{A yearmon object specifying the month to check for overlapping events}
}
\value{
An integer representing the number of events overlapping with the specified month
}
\description{
Count Events Overlapping with a Specific Month
}
\examples{
num_events <- count_events_in_month(as.yearmon("Jan 2003"))
print(paste("Number of overlapping events:", num_events))

}
