% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{president}
\alias{president}
\title{Helper variable for levels of presidential administration}
\format{
A list with members `president$levels` (a chronological list of
names for each presidency in English), `president$initials` (two-
or three-letter initials for display in shortened contexts).
}
\usage{
president
}
\description{
Helper variable for levels of presidential administration
}
\keyword{datasets}
