% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel-pres-admin.R
\name{relabel_pres_admin}
\alias{relabel_pres_admin}
\title{Relabel presidential administration using presidency_name_table}
\usage{
relabel_pres_admin(
  dataframe,
  output_column,
  .input_column = "presidency",
  keep_original = TRUE
)
}
\arguments{
\item{dataframe}{The input dataframe containing a pres_admin column}

\item{output_column}{The column from presidency_name_table to use for new labels
(unquoted column name)}

\item{keep_original}{Logical indicating whether to keep the original pres_admin
column as pres_admin_original (default TRUE)}
}
\value{
The dataframe with pres_admin relabeled according to output_column
}
\description{
Replaces pres_admin values with corresponding values from a specified column
in presidency_name_table (e.g., presidency_year, presidency_short, etc.)
}
\examples{
def_labeled <- relabel_pres_admin(def, presidency_year)
def_labeled <- relabel_pres_admin(def, presidency_short)
def_labeled <- relabel_pres_admin(def, presidency_full, keep_original = FALSE)
}
