% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-outcome-reactable.R
\name{wide_event_outcomes_reactable}
\alias{wide_event_outcomes_reactable}
\title{Display the Detailed Event Outcomes Table as an Interactive Reactable}
\usage{
wide_event_outcomes_reactable(dataset, static = FALSE, max_larger = 0)
}
\arguments{
\item{dataset}{A data frame in the format generated by `wide_event_outcomes_table()`,
possibly with extra columns added}

\item{static}{Logical, whether the table should be static or interactive (default: TRUE)}

\item{max_larger}{Numeric, allows setting a custom maximum value for death counts (default: 0)}
}
\value{
A reactable object displaying the event outcomes table
}
\description{
Creates an interactive table to display event, death total, and outcome
information. Columns counting deaths are group under a "deaths" heading.
Outcomes are color coded.
}
\examples{
deaths_aug24_filtered <- standard_filter(deaths_aug24)
deaths_aug24_unfiltered <- deaths_aug24
wetable <- wide_event_outcomes_table(deaths_aug24_filtered,
  deaths_aug24_unfiltered, event_status_aug24, drop_extra=TRUE, drop_separate=TRUE)
wide_event_outcomes_reactable(wetable)
}
